Sample Changes to implement "RunSequence" function
Demonstrates an implementation of a function that runs each model and displays data as it runs. Essentially does the old work of agents.py in a single function of the agent.
NOTE: STILL NEEDS TO BE ADDED FOR OTHER AGENTS THAN JUST EPSILON GREEDY. This is just the work I have completed as of right now, and is not fully finished